Investment Opportunities Screen completion
