--- components/ProgressBar.js
@@ -1,8 +1,0 @@
-import * as React from 'react';
-import { Text } from 'react-native';
-
-export function ProgressBar() {
-  return (
-      <Text> This is the ProgressBar component</Text>
-  );
-}
