+++ screens/OnboardFourScreen.js
@@ -63,6 +63,7 @@ const styles = StyleSheet.create({
     flex: 1,
     flexDirection: "column",
     justifyContent: "space-between",
+    // justifyContent: "center",
     marginTop: 50,
   },
   headerText: {
